+++ dev/py-test.sh
@@ -11,5 +11,4 @@ cd "$FWDIR"
 pushd python
 pytest -s -m spark
 pytest -s -m hail
-pytest -s -m covariates
 popd
