+++ python/requirements.txt
@@ -4,3 +4,4 @@ pandas==1.1.4
 typedecorator==0.0.5
 hail==0.2.74
 pyspark==3.1.2
+Jinja2==3.0.3
