+++ python/varspark/test/test_hail.py
@@ -1,6 +1,7 @@
 import filecmp
 import os
 import unittest
+from io import StringIO
 
 import hail as hl
 import numpy.testing as npt
@@ -69,6 +70,45 @@ class RFModelHailTest(unittest.TestCase):
                 os.path.join(PROJECT_DIR, 'target/test-json-model.json'))
             npt.assert_approx_equal(expected_oob_error, rf_model.oob_error())
             pdt.assert_frame_equal(expected_top100_imp, actual_top100_imp)
+
+
+
+    def test_covariates(self):
+
+
+        data = hl.import_vcf(os.path.join(PROJECT_DIR, 'data/chr22_1000.vcf'))
+        labels = hl.import_table(os.path.join(PROJECT_DIR, 'data/chr22-labels.csv'),
+                                 delimiter=',',
+                                 types={'22_16050408': 'float64'}).key_by('sample')
+
+        # loading the covariates and transposing them
+        covariates = pd.read_csv(
+            os.path.join(PROJECT_DIR, 'data/chr22_1000_pheno-wide.csv')).T.reset_index()
+        covariates.at[0, 'index'] = 'samples'
+        covariates.columns = covariates.iloc[0]
+        covariates = covariates.iloc[1:, :]
+        covariates = hl.Table.from_pandas(covariates).key_by('samples')
+        mt = data.annotate_cols(covariates=covariates[data.s],
+                                labels=labels.select('22_16050408')[data.s])
+
+        rf_model = vshl.random_forest_model(y=mt.labels['22_16050408'],
+                                                  x=mt.GT.n_alt_alleles(),
+                                                  covariates={'age': mt.covariates.age,
+                                                              'bmi': mt.covariates.bmi},
+                                                  seed=13, mtry_fraction=0.05,
+                                                  min_node_size=5, max_depth=10)
+
+        rf_model.fit_trees(100, 50)
+
+
+        impTable = rf_model.variable_importance()
+        self.assertEqual(len(impTable.to_pandas()), 1988)
+
+        covImpTable = rf_model.covariate_importance()
+        self.assertEqual(len(covImpTable.to_pandas()), 2)
+
+
+
 
 
 if __name__ == '__main__':
